+++ oneflow/ir/lib/OneFlow/OneFlowOps.cpp
@@ -263,68 +263,6 @@ void AddNOp::getCanonicalizationPatterns(::mlir::RewritePatternSet& results,
   results.insert<ConvertAddOpWithArity>(context);
 }
 
-// TODO: merge all ctrl input and output when folding op
-bool HaveIdenticalPlacement(mlir::Operation* a, mlir::Operation* b) {
-  UserOpAdaptor adaptor_a(a->getOperands(), a->getAttrDictionary());
-  UserOpAdaptor adaptor_b(b->getOperands(), b->getAttrDictionary());
-  return adaptor_a.device_tag() == adaptor_b.device_tag()
-         && adaptor_a.device_name() == adaptor_b.device_name();
-}
-
-using namespace OpTrait;
-OpFoldResult OpTrait::impl::foldIdempotentOfIdenticalPlacement(Operation* op) {
-  auto* argument_op = op->getOperand(0).getDefiningOp();
-  if (argument_op && op->getName() == argument_op->getName()
-      && HaveIdenticalPlacement(op, argument_op)) {
-    return op->getOperand(0);
-  }
-  return {};
-}
-
-OpFoldResult OpTrait::impl::foldInvolutionOfIdenticalPlacement(Operation* op) {
-  auto* argument_op = op->getOperand(0).getDefiningOp();
-  if (argument_op && op->getName() == argument_op->getName()
-      && HaveIdenticalPlacement(op, argument_op)) {
-    return argument_op->getOperand(0);
-  }
-  return {};
-}
-
-LogicalResult OpTrait::impl::VerifyIsOpConfCompatible(Operation* op) {
-  for (auto attr : {
-           IsOpConfCompatible<void>::getOpNameAttr(),
-           IsOpConfCompatible<void>::getDeviceTagAttr(),
-       }) {
-    if (!op->hasAttrOfType<StringAttr>(attr)) {
-      return op->emitError("expected operation to have attribute: " + attr);
-    }
-  }
-  if (!op->hasAttrOfType<ArrayAttr>(IsOpConfCompatible<void>::getDeviceNameAttr())) {
-    return op->emitError("expected operation to have attribute: "
-                         + IsOpConfCompatible<void>::getDeviceNameAttr());
-  }
-  return success();
-}
-
-LogicalResult OpTrait::impl::VerifyIsImportCompatible(Operation* op) {
-  if (auto output_lbns =
-          op->getAttrOfType<ArrayAttr>(IsImportCompatible<void>::getOutputLBNsAttr())) {
-    if (auto cec = dyn_cast<ControlEdgeCompatible>(op)) {
-      if (cec.dataOutputResults().size() != output_lbns.size()) {
-        return op->emitError("expected number of data output results to be "
-                             + std::to_string(output_lbns.size()) + " but got "
-                             + std::to_string(cec.dataOutputResults().size()));
-      }
-    } else {
-      return op->emitError("expected to support ControlEdgeCompatible");
-    }
-  } else {
-    return op->emitError("expected operation to have attribute: "
-                         + IsImportCompatible<void>::getOutputLBNsAttr());
-  }
-  return success();
-}
-
 void NormalizationAddReluOp::build(::mlir::OpBuilder& odsBuilder, ::mlir::OperationState& odsState,
                                    Value x, Value addend, Value moving_mean, Value moving_variance,
                                    Value gamma, Value beta, StringRef op_name, StringRef device_tag,
